+++ packages/generator-component/generators/app/templates/stories/Skeleton.__stories__.js
@@ -2,8 +2,14 @@
 // import {
 //     withKnobs, select, boolean
 // } from '@storybook/addon-knobs';
+import { select } from '@storybook/addon-knobs';
 import { withA11y } from '@storybook/addon-a11y';
 import <%= name.component %> from '../src/components/<%= name.filename %>.vue';
+
+import {
+    VALID_LOCALES,
+    ENGLISH_LOCALE
+} from '../../storybook/constants/globalisation';
 
 export default {
     title: 'Components',
@@ -13,14 +19,11 @@ export default {
 export const <%= name.component %>Component = () => ({
     components: { <%= name.component %> },
     props: {
-        buttonType: {
-            default: select('Button Type', ['primary', 'primaryAlt', 'secondary', 'tertiary', 'link'])
-        },
-        fullWidth: {
-            default: boolean('fullWidth', false)
+        locale: {
+            default: select('Locale', VALID_LOCALES, ENGLISH_LOCALE)
         }
     },
-    template: `<<%= name.template %> />`
+    template: `<<%= name.template %> :locale="locale" />`
 });
 
 <%= name.component %>Component.storyName = 'f-<%= name.default %>';