+++ packages/generator-component/CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v1.9.0
+------------------------------
+*December 07, 2020*
+
+### Added
+- F-Globalisation wiring to the component template
+
+### Removed
+- Erroneous props in the story template
+
 
 v1.8.0
 ------------------------------